refactor: перешел на ts-node, сделал все модульным

+++ src/modules/users/listeners/MemberUpdate.ts
@@ -1,14 +1,15 @@
 import { Events, Listener } from '@sapphire/framework';
 import { ApplyOptions } from '@sapphire/decorators';
-import type { GuildMember, Role } from 'discord.js';
-import { UserRoles } from '../lib/directus/directus-entities/Discord';
-import { FilterRule } from '../lib/directus/directus-orm/filters';
+import type { GuildMember } from 'discord.js';
+import { UserRoles } from '@/lib/directus/directus-entities/Discord';
+import { FilterRule } from '@/lib/directus/directus-orm/filters';
 
 @ApplyOptions<Listener.Options>({ event: Events.GuildMemberUpdate })
 export class MemberUpdate extends Listener<typeof Events.GuildMemberUpdate> {
   async run(oldMember: GuildMember, newMember: GuildMember) {
     const deletedRoles: string[] = [];
     const addedRoles: string[] = [];
+
     if (oldMember.roles.cache.size > newMember.roles.cache.size) {
       for (const [role] of oldMember.roles.cache) {
         if (!newMember.roles.cache.has(role)) {
@@ -32,9 +33,7 @@ export class MemberUpdate extends Listener<typeof Events.GuildMemberUpdate> {
       const findRole = userRoles.find(
         (userRole) => userRole.Discord_Roles_id === role,
       );
-      if (findRole) {
-        await findRole.delete();
-      }
+      if (findRole) await findRole.delete();
     }
 
     for (const role of addedRoles) {
